Rails upgrade from 6.1 to 8.0  (#528)
* upgrade outdated gems

* Upgrade sprockets

* Upgrade rails version to 7.0

* Fix base64 version

* Fix round players query, upgrade yarn libraries

* Add default_url_options for routes

* Upgrade Rails gem to version 7.1

* Fix ci config

* Disable http cache

* Upgrade haml gem

* Add missed_players_data for national matches tour (#535)

* Upgrade rails to 7.2 version

* Fix lineups and matches order on fanta tour page

* Upgrade rails to version 8.0

* Fix club transfers rake

+++ app/services/players/query.rb
@@ -100,8 +100,8 @@ module Players
     end
 
     def filter_by_appearances(players)
-      players = players.where('COALESCE(pss.played_matches, 0) >= ?', min_app.to_f) if min_app
-      players = players.where('COALESCE(pss.played_matches, 0) <= ?', max_app.to_f) if max_app
+      players = players.where('COALESCE(pss.played_matches, 0) >= ?', min_app.to_i) if min_app
+      players = players.where('COALESCE(pss.played_matches, 0) <= ?', max_app.to_i) if max_app
       players
     end
